requirements.txt
ag2==0.8.3
annotated-types==0.7.0
anyio==4.9.0
asyncer==0.0.8
Authlib==1.3.1
autogen==0.8.3
beautifulsoup4==4.13.3
blis==1.2.0
bump-pydantic==0.8.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.11
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
kagglehub==0.3.10
langcodes==3.5.0
language_data==1.3.0
libcst==1.7.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.12
numpy==2.2.4
openai==1.68.2
packaging==24.2
preshed==3.0.9
protobuf==5.29.4
pyautogen==0.8.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
pywin32==310
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
setuptools==77.0.3
shellingham==1.5.4
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.46.1
tenacity==9.0.0
termcolor==2.5.0
thinc==8.3.4
tiktoken==0.9.0
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
validators==0.34.0
wasabi==1.1.3
weasel==0.4.1
weaviate-client==4.11.2
wrapt==1.17.2
